+++ library/src/ipfs_client/ipld/dag_headers.cc
@@ -1,23 +1,12 @@
 #include <ipfs_client/ipld/dag_headers.h>
 
 #include <chrono>
-#include <chrono>
-#include <chrono>
-#include <chrono>
 #include <sstream>
 #include <string_view>
 #include <string>
 
 #include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
-#include "ipfs_client/gw/gateway_request_type.h"
 #include "ipfs_client/ipld/block_source.h"
-#include "ipfs_client/gw/gateway_request_type.h"
 
 using Self = ipfs::ipld::DagHeaders;
 using ReqTyp = ipfs::gw::GatewayRequestType;
@@ -64,14 +53,14 @@ void Self::Add(BlockSource const& src) {
     value << "\";dur="
           << c::duration_cast<c::milliseconds>(src.load_duration).count();
     headers_.emplace_back("Server-Timing", value.str());
-    if (!src.cid.empty() != 0u) {
+    if (static_cast<unsigned int>(!src.cid.empty()) != 0U) {
       auto from = src.cat.cached ? std::string{"cache"} : src.cat.gateway_url;
       headers_.emplace_back("IPFS-Source-" + src.cid, from);
     }
   }
 }
 void Self::Finish() {
-  if (other_count_ == 0u) {
+  if (other_count_ == 0U) {
     return;
   }
   std::ostringstream value;
